factorize numba into the evaluator
